feat: refactor general config page + add LLM conf warning
