+++ components/Form/ProvidersConfig.vue
@@ -10,23 +10,10 @@
   </div>
 </template>
 <script lang="ts" setup>
-
-definePageMeta({
-  middleware: [
-    function () {
-      const statesStore = useStatesStore();
-      const isAdmin = statesStore.userHasRole('admin');
-      if (!isAdmin) {
-        return navigateTo('/');
-      }
-    },
-  ],
-});
-
 const loadProvidersList = async () => {
   const items = await $fetch('/api/brevia/providers?list_models=false');
   return items.map((p: any) => String(p.model_provider));
-}
+};
 
 const providersList: string[] = await loadProvidersList();
 const store = useProvidersStore();
@@ -37,7 +24,7 @@ const loadConfigData = async () => {
   store.providers = breviaConfig.providers;
   store.setKnownVars(breviaConfig.providers_env_vars);
   store.setupEnvVarsMap(breviaConfig.brevia_env_secrets);
-}
+};
 
 await loadConfigData();
 </script>
