+++ locales/en/messages.json
@@ -86,6 +86,7 @@
     "FIRST_NAME_PLACEHOLDER": "Your first name",
     "FIRST_NAME_ERROR": "The user name cannot be empty!",
     "FORGOT_PASS": "Forgot password?",
+    "GENERAL": "General",
     "GIVE_THE_NEW_CHATBOT_A_TITLE_AND_INCLUDE_A_BRIEF_DESCRIPTION": "Give the new chatbot a title and include a brief description",
     "GO_TO_LOGIN_PAGE": "Go to login page",
     "HELLO": "Hello {name}",
@@ -108,6 +109,8 @@
     },
     "LINK_PLACEHOLDER": "Write here a valid link to a webpage",
     "LINKS": "Links",
+    "LLM_CONF_UNSUPPORTED": "Unsupported LLM configuration",
+    "LLM_CONF_UNSUPPORTED_DESCRIPTION": "You are using an LLM configuration that is not supported by the current settings. In this mode you can change it according to the settings.",
     "LOGIN_PLACEHOLDER": "Enter your email or username",
     "LOGOUT": "Logout",
     "MAIL_TO": "An email has been sent to ",
@@ -142,6 +145,7 @@
     "PASSWORD_PLACEHOLDER": "Enter password",
     "PASSWORD_RESET": "Reset your password",
     "PERMANENTLY": "permanently",
+    "PROVIDERS": "Providers",
     "QUESTION": "Question",
     "QUESTION_PLACEHOLDER": "Write here a potential question to guide the chatbot's behaviour",
     "RECAPTCHA": "This form is protected by reCAPTCHA and Google's {privacyPolicy} and {termsOfService} apply",
