+++ locales/it/messages.json
@@ -86,6 +86,7 @@
     "FIRST_NAME_PLACEHOLDER": "Il tuo nome",
     "FIRST_NAME_ERROR": "Il nome utente non può essere vuoto!",
     "FORGOT_PASS": "Password dimenticata?",
+    "GENERAL": "Generale",
     "GIVE_THE_NEW_CHATBOT_A_TITLE_AND_INCLUDE_A_BRIEF_DESCRIPTION": "Dai un titolo al nuovo chatbot e inserisci una breve descrizione",
     "GO_TO_LOGIN_PAGE": "Vai alla pagina di accesso",
     "LAST_LOGIN_ERROR": "Ultimo errore di login: ",
@@ -101,6 +102,8 @@
     },
     "LINK_PLACEHOLDER": "Scrivi qui un link valido a una pagina web",
     "LINKS": "Link",
+    "LLM_CONF_UNSUPPORTED": "Configurazion LLM non supportata",
+    "LLM_CONF_UNSUPPORTED_DESCRIPTION": "Stai utilizzando una configurazione LLM non supportata dalle attuali impostazioni. In questa modalità puoi modificarla secondo le impostazioni.",
     "HELLO": "Ciao {name}",
     "INDEXING_SUCCESS": "Contenuto indicizzato correttamente!",
     "INDEXING_FAILURE": "Si è verificato un errore durante l'indicizzazione. Riprova più tardi.",
@@ -142,6 +145,7 @@
     "PASSWORD_PLACEHOLDER": "Inserisci password",
     "PASSWORD_RESET": "Reimposta la password",
     "PERMANENTLY": "definitivamente",
+    "PROVIDERS": "Provider",
     "PLEASE_DROP_A_FILE": "Trascina qui un file o clicca per selezionarlo",
     "PRIVACY_POLICY": "Privacy Policy",
     "QUESTION": "Domanda",
